fix: move buildSchemaSyncPayload to @mdcms/shared/server to avoid node:crypto in browser bundle
